requirements.txt
streamlit
pillow
qrcode[pil]
requests